+++ metadata.rb
@@ -4,12 +4,12 @@ maintainer_email 'mark@markcallen.com'
 license          'All rights reserved'
 description      'Installs/Configures chef11server on uhost'
 long_description 'Installs/Configures chef11server on uhost'
-version          '0.1.0'
+version          '0.2.0'
 
-%w{ ubuntu }.each do |os|
+%w{ ubuntu fedora centos }.each do |os|
   supports os
 end
 
-%w{ apt nginx hostsfile }.each do |cb|
+%w{ apt nginx hostsfile users }.each do |cb|
   depends cb
 end