+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0 (Nov 13, 2014)
+
+Features:
+  - Creates uhost user
+  - Installs uhostappserver and proxies with nginx on port 443
+
+Bugfixes: 
+  - none
+
 ## 0.1.0 (May 9, 2014)
 
 Features: